feat: first bit of abstraction
