+++ src/Components/ScheduleEditor/ScheduleEditor.module.scss
@@ -1,0 +1,6 @@
+.modalContainer {
+  height: 52rem;
+  max-width: 50rem;
+  width: 50rem;
+  max-height: 84vh;
+}
