+++ src/Components/ScheduleEditor/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./ScheduleEditor";
