--- src/Features/Editor/Configure/CronJobModal/index.tsx
@@ -1,1 +1,0 @@
-export { default } from "./CronJobModal";
