--- src/Features/Editor/Schedule/CronJobConfig/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './CronJobConfig';
