--- src/Features/Editor/Schedule/RunOnceConfig/RunOnceConfig.tsx
@@ -1,10 +1,0 @@
-import React from "react";
-import { ScheduleUnion } from "Types";
-
-interface RunOnceConfigProps {
-  event: ScheduleUnion;
-}
-
-export default function RunOnceConfig(props: RunOnceConfigProps) {
-  return <div>Run once!</div>;
-}
