--- src/Features/Editor/Schedule/RunOnceConfig/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './RunOnceConfig';
