+++ src/Features/Editor/Schedule/Schedule.module.scss
@@ -160,10 +160,3 @@
     @include text-overflow-ellipsis();
   }
 }
-
-.modalContainer {
-  height: 52rem;
-  max-width: 50rem;
-  width: 50rem;
-  max-height: 84vh;
-}
