+++ src/Features/Editor/Schedule/Schedule.tsx
@@ -3,58 +3,33 @@ import { useQuery, useMutation, queryCache, QueryResult } from "react-query";
 import {
   Button,
   ConfirmModal,
-  ComposedModal,
   CodeSnippet,
-  Creatable,
-  ComboBox,
-  DynamicFormik,
-  ModalBody,
-  ModalForm,
-  ModalFooter,
   MultiSelect,
   Error,
   Loading,
   OverflowMenu,
   OverflowMenuItem,
-  RadioButtonGroup,
-  RadioButton,
   Search,
   Tag,
-  TextArea,
-  TextInput,
   Tile,
   TooltipHover,
   ToastNotification,
   notify,
 } from "@boomerang-io/carbon-addons-boomerang-react";
-import CronJobConfig from "./CronJobConfig";
+import ScheduleCreator from "Components/ScheduleCreator";
+import ScheduleEditor from "Components/ScheduleEditor";
 import SlidingPane from "react-sliding-pane";
 import queryString from "query-string";
 import Calendar from "Components/Calendar";
 import cronstrue from "cronstrue";
 import matchSorter from "match-sorter";
 import moment from "moment-timezone";
-import * as Yup from "yup";
-import { cronToDateTime, cronDayNumberMap } from "Utils/cronHelper";
-import {
-  DATETIME_LOCAL_DISPLAY_FORMAT,
-  DATETIME_LOCAL_INPUT_FORMAT,
-  defaultTimeZone,
-  timezoneOptions,
-  transformTimeZone,
-} from "Utils/dateHelper";
+import { DATETIME_LOCAL_DISPLAY_FORMAT } from "Utils/dateHelper";
 import { scheduleStatusOptions, statusLabelMap, typeLabelMap } from "Features/Schedule";
 import { queryStringOptions } from "Config/appConfig";
 import { serviceUrl, resolver } from "Config/servicesConfig";
 import { Add16, CircleFilled16, SettingsAdjust16, RadioButton16, Repeat16, RepeatOne16 } from "@carbon/icons-react";
-import {
-  CalendarEvent,
-  ComposedModalChildProps,
-  ScheduleCalendar,
-  ScheduleType,
-  ScheduleUnion,
-  WorkflowSummary,
-} from "Types";
+import { CalendarEvent, ScheduleCalendar, ScheduleUnion, WorkflowSummary } from "Types";
 import "react-sliding-pane/dist/react-sliding-pane.css";
 import styles from "./Schedule.module.scss";
 
@@ -156,7 +131,7 @@ export default function ScheduleView(props: ScheduleProps) {
         setIsOpen={setIsPanelOpen}
         setIsEditorOpen={setIsEditorOpen}
       />
-      <CreateSchedule
+      <ScheduleCreator
         isModalOpen={isCreatorOpen}
         onCloseModal={() => setIsCreatorOpen(false)}
         schedule={activeSchedule}
@@ -164,7 +139,7 @@ export default function ScheduleView(props: ScheduleProps) {
         workflowScheduleUrl={workflowScheduleUrl}
         workflowCalendarUrl={workflowCalendarUrl}
       />
-      <EditSchedule
+      <ScheduleEditor
         isModalOpen={isEditorOpen}
         onCloseModal={() => setIsEditorOpen(false)}
         schedule={activeSchedule}
@@ -430,7 +405,8 @@ function ScheduledListItem(props: ScheduledListItemProps) {
       },
     },
     {
-      itemText: isActive ? "Disable" : "Enable",
+      itemText: props.schedule.status === "inactive" ? "Enable" : "Disable",
+      disabled: props.schedule.status === "trigger_disabled",
       onClick: () => setIsToggleStatusModalOpen(true),
     },
     {
@@ -630,456 +606,5 @@ function EditorPanel(props: PanelProps) {
     >
       {renderSchedule()}
     </SlidingPane>
-  );
-}
-
-/**
- * Start the beast of a create schedule form
- */
-interface CreateScheduleProps {
-  isModalOpen: boolean;
-  onCloseModal: () => void;
-  schedule?: ScheduleUnion;
-  workflow: WorkflowSummary;
-  workflowScheduleUrl: string;
-  workflowCalendarUrl: string;
-}
-
-interface CreateScheduleForm {
-  id: string;
-  name: string;
-  description: string;
-  cronSchedule: string;
-  dateTime: string;
-  labels: Array<string>;
-  type: ScheduleType;
-  days: Array<string>;
-  timezone: { label: string; value: string };
-  time: string;
-  advancedCron: boolean;
-  parameters: { [key: string]: any };
-}
-
-function CreateSchedule(props: CreateScheduleProps) {
-  /**
-   * Create schedule
-   */
-  const [createScheduleMutator, createScheduleMutation] = useMutation(resolver.postSchedule, {});
-
-  const handleCreateSchedule = async (schedule: ScheduleUnion) => {
-    try {
-      await createScheduleMutator({ body: schedule });
-      notify(
-        <ToastNotification
-          kind="success"
-          title={`Created Schedule`}
-          subtitle={`Successfully created schedule ${schedule.name} `}
-        />
-      );
-      queryCache.invalidateQueries(props.workflowScheduleUrl);
-      queryCache.invalidateQueries(props.workflowCalendarUrl);
-    } catch (e) {
-      notify(
-        <ToastNotification
-          kind="error"
-          title="Something's Wrong"
-          subtitle={`Request to create schedule ${schedule.name} failed`}
-        />
-      );
-      return;
-    }
-  };
-
-  const handleSubmit = async (values: CreateScheduleForm) => {
-    const { name, description, cronSchedule, dateTime, labels, timezone, type, days, time, ...parameters } = values;
-
-    let scheduleLabels: Array<{ key: string; value: string }> = [];
-    if (values.labels.length) {
-      scheduleLabels = values.labels.map((pair: string) => {
-        const [key, value] = pair.split(":");
-        return { key, value };
-      });
-    }
-    let scheduleType = type;
-    const schedule: Partial<ScheduleUnion> = {
-      name,
-      description,
-      type: scheduleType,
-      timezone: timezone.value,
-      labels: scheduleLabels,
-      parameters,
-      workflowId: props.workflow.id,
-    };
-
-    if (schedule.type === "runOnce") {
-      schedule["dateSchedule"] = new Date(dateTime).toISOString();
-    }
-
-    if (schedule.type === "cron") {
-      let daysCron: Array<string> | [] = [];
-      Object.values(days).forEach((day) => {
-        //@ts-ignore
-        daysCron.push(cronDayNumberMap[day]);
-      });
-      const timeCron = !time ? ["0", "0"] : time.split(":");
-      const cronSchedule = `0 ${timeCron[1]} ${timeCron[0]} ? * ${daysCron.length !== 0 ? daysCron.toString() : "*"}`;
-      schedule["cronSchedule"] = cronSchedule;
-    }
-
-    if (schedule.type === "advancedCron") {
-      schedule["cronSchedule"] = cronSchedule;
-    }
-
-    await handleCreateSchedule(schedule as ScheduleUnion);
-  };
-
-  return (
-    <ComposedModal
-      isOpen={props.isModalOpen}
-      onCloseModal={props.onCloseModal}
-      composedModalProps={{
-        containerClassName: styles.modalContainer,
-      }}
-      modalHeaderProps={{
-        title: "Create a Schedule",
-      }}
-    >
-      {(modalProps: ComposedModalChildProps) => (
-        <CreateEditForm
-          modalProps={modalProps}
-          isLoading={createScheduleMutation.isLoading}
-          handleSubmit={handleSubmit}
-          //@ts-ignore
-          parameters={props.workflow.properties}
-          schedule={props.schedule}
-          type="create"
-        />
-      )}
-    </ComposedModal>
-  );
-}
-
-interface EditScheduleProps {
-  isModalOpen: boolean;
-  onCloseModal: () => void;
-  schedule?: ScheduleUnion;
-  workflowScheduleUrl: string;
-  workflowCalendarUrl: string;
-  workflow: WorkflowSummary;
-}
-
-function EditSchedule(props: EditScheduleProps) {
-  /**
-   * Update schedule
-   */
-  const [updateScheduleMutator, { isLoading: updateScheduleIsLoading }] = useMutation(resolver.patchSchedule, {});
-
-  const handleUpdateSchedule = async (updatedSchedule: ScheduleUnion) => {
-    if (props.schedule) {
-      try {
-        await updateScheduleMutator({ body: updatedSchedule, scheduleId: props.schedule.id });
-        notify(
-          <ToastNotification
-            kind="success"
-            title={`Updated Schedule`}
-            subtitle={`Successfully updated schedule ${props.schedule.name} `}
-          />
-        );
-        queryCache.invalidateQueries(props.workflowScheduleUrl);
-        queryCache.invalidateQueries(props.workflowCalendarUrl);
-      } catch (e) {
-        notify(
-          <ToastNotification
-            kind="error"
-            title="Something's Wrong"
-            subtitle={`Request to update schedule ${props.schedule.name} failed`}
-          />
-        );
-        return;
-      }
-    }
-  };
-
-  const handleSubmit = async (values: CreateScheduleForm) => {
-    const { id, name, description, cronSchedule, dateTime, labels, timezone, type, days, time, ...parameters } = values;
-
-    let scheduleLabels: Array<{ key: string; value: string }> = [];
-    if (values.labels.length) {
-      scheduleLabels = values.labels.map((pair: string) => {
-        const [key, value] = pair.split(":");
-        return { key, value };
-      });
-    }
-    let scheduleType = type;
-    const schedule: Partial<ScheduleUnion> = {
-      name,
-      description,
-      type: scheduleType,
-      timezone: timezone.value,
-      labels: scheduleLabels,
-      parameters,
-    };
-
-    if (schedule.type === "runOnce") {
-      schedule["dateSchedule"] = new Date(dateTime).toISOString();
-    }
-
-    if (schedule.type === "cron") {
-      let daysCron: Array<string> | [] = [];
-      Object.values(days).forEach((day) => {
-        //@ts-ignore
-        daysCron.push(cronDayNumberMap[day]);
-      });
-      const timeCron = !time ? ["0", "0"] : time.split(":");
-      const cronSchedule = `0 ${timeCron[1]} ${timeCron[0]} ? * ${daysCron.length !== 0 ? daysCron.toString() : "*"}`;
-      schedule["cronSchedule"] = cronSchedule;
-    }
-
-    if (schedule.type === "advancedCron") {
-      schedule["cronSchedule"] = cronSchedule;
-    }
-
-    await handleUpdateSchedule(schedule as ScheduleUnion);
-  };
-
-  return (
-    <ComposedModal
-      isOpen={props.isModalOpen}
-      onCloseModal={props.onCloseModal}
-      composedModalProps={{
-        containerClassName: styles.modalContainer,
-      }}
-      modalHeaderProps={{
-        title: "Edit a Schedule",
-      }}
-    >
-      {(modalProps: ComposedModalChildProps) => (
-        <CreateEditForm
-          isLoading={updateScheduleIsLoading}
-          handleSubmit={handleSubmit}
-          modalProps={modalProps}
-          parameters={props.workflow.properties}
-          schedule={props.schedule}
-          type={"edit"}
-        />
-      )}
-    </ComposedModal>
-  );
-}
-
-interface CreateEditFormProps {
-  handleSubmit: (args: CreateScheduleForm) => void;
-  isLoading: boolean;
-  modalProps: ComposedModalChildProps;
-  parameters?: { [k: string]: any };
-  type: "create" | "edit";
-  schedule?: ScheduleUnion;
-}
-
-function CreateEditForm(props: CreateEditFormProps) {
-  let initFormValues: Partial<CreateScheduleForm> = {
-    id: props.schedule?.name,
-    name: props.schedule?.name ?? "",
-    description: props.schedule?.description ?? "",
-    type: props.schedule?.type || "runOnce",
-    timezone: transformTimeZone(defaultTimeZone),
-    days: [],
-    labels: [],
-    ...props.schedule?.parameters,
-  };
-
-  /**
-   * Handle creating it from calendar click
-   */
-  if (props.type === "create" && props.schedule && props.schedule.type === "runOnce") {
-    initFormValues["dateTime"] = moment(props.schedule.dateSchedule).format(DATETIME_LOCAL_INPUT_FORMAT);
-  }
-
-  /**
-   * Lots of manipulating of data for the inputs based on type
-   */
-  if (props.type === "edit" && props.schedule) {
-    initFormValues["timezone"] = transformTimeZone(props.schedule.timezone);
-
-    if (props.schedule.type === "runOnce") {
-      initFormValues["dateTime"] = moment(props.schedule.dateSchedule).format(DATETIME_LOCAL_INPUT_FORMAT);
-    }
-
-    if (props.schedule.type === "advancedCron") {
-      initFormValues["cronSchedule"] = props.schedule.cronSchedule;
-    }
-
-    if (props.schedule.type === "cron") {
-      const cronSchedule = props.schedule.cronSchedule;
-      const cronToData = cronToDateTime(Boolean(cronSchedule), cronSchedule);
-      const { cronTime, selectedDays } = cronToData;
-
-      let activeDays: string[] = [];
-      Object.entries(selectedDays).forEach(([key, value]) => {
-        if (value) {
-          activeDays.push(key);
-        }
-      });
-
-      initFormValues["time"] = cronTime;
-      initFormValues["days"] = activeDays;
-    }
-
-    let scheduleLabels: Array<string> = [];
-    if (props.schedule.labels?.length) {
-      for (let labelObj of props.schedule.labels) {
-        const scheduleLabel = `${labelObj.key}:${labelObj.value}`;
-        scheduleLabels.push(scheduleLabel);
-      }
-      initFormValues["labels"] = scheduleLabels;
-    }
-  }
-
-  return (
-    <DynamicFormik
-      validateOnMount
-      initialValues={initFormValues}
-      inputs={props.parameters}
-      onSubmit={async (args: CreateScheduleForm) => {
-        await props.handleSubmit(args);
-        props.modalProps.closeModal();
-      }}
-      validationSchemaExtension={Yup.object().shape({
-        name: Yup.string().required("Name is required"),
-        description: Yup.string(),
-        type: Yup.string().required("Enter a type"),
-        dateTime: Yup.string()
-          .when("type", {
-            is: "runOnce",
-            then: Yup.string().required("Date and Time is required"),
-          })
-          .test("isAfterNow", "Enter a date and time after now", (value: string | undefined) => {
-            return moment(value).isAfter(new Date());
-          }),
-        labels: Yup.array(),
-        cronSchedule: Yup.string().when("type", {
-          is: "advancedCron",
-          then: Yup.string().required("Expression required"),
-        }),
-        days: Yup.array().when("type", {
-          is: "cron",
-          then: Yup.array().min(1, "Enter at least one day"),
-        }),
-        time: Yup.string().when("type", { is: "cron", then: (time: any) => time.required("Enter a time") }),
-        timezone: Yup.object().shape({ label: Yup.string(), value: Yup.string() }),
-      })}
-    >
-      {({ inputs, formikProps }: any) => (
-        <ModalForm noValidate onSubmit={formikProps.handleSubmit}>
-          <ModalBody>
-            <p>
-              <b>About</b>
-            </p>
-            <TextInput
-              labelText="Name"
-              id="name"
-              onBlur={formikProps.handleBlur}
-              onChange={formikProps.handleChange}
-              invalid={formikProps.errors.name && formikProps.touched.name}
-              invalidText={formikProps.errors.name}
-              placeholder="e.g. Daily task"
-              value={formikProps.values.name}
-            />
-            <TextArea
-              labelText="Description (optional)"
-              id="description"
-              placeholder="e.g. Runs very important daily task."
-              onBlur={formikProps.handleBlur}
-              onChange={formikProps.handleChange}
-              invalid={formikProps.errors.description && formikProps.touched.description}
-              invalidText={formikProps.errors.description}
-              value={formikProps.values.description}
-            />
-            <Creatable
-              createKeyValuePair
-              keyLabelText="Label key"
-              keyPlaceholder="level"
-              valueLabelText="Label value"
-              valuePlaceholder="important"
-              value={formikProps.values.labels}
-              onChange={(labels: string) => formikProps.setFieldValue("labels", labels)}
-            />
-            <p>
-              <b>Schedule</b>
-            </p>
-            <section>
-              <p>What type of Schedule do you want to create?</p>
-              <RadioButtonGroup
-                id="type"
-                labelPosition="right"
-                name="type"
-                onChange={(type: string) => formikProps.setFieldValue("type", type)}
-                orientation="horizontal"
-                valueSelected={formikProps.values["type"]}
-              >
-                <RadioButton key={"runOnce"} id={"runOnce"} labelText={typeLabelMap["runOnce"]} value={"runOnce"} />
-                <RadioButton key={"cron"} id={"cron"} labelText={typeLabelMap["cron"]} value={"cron"} />
-                <RadioButton
-                  key={"advanced-cron"}
-                  id={"advanced-cron"}
-                  labelText={typeLabelMap["advancedCron"]}
-                  value={"advancedCron"}
-                />
-              </RadioButtonGroup>
-            </section>
-            {formikProps.values["type"] === "runOnce" ? (
-              <>
-                <div style={{ width: "23.5rem" }}>
-                  <TextInput
-                    type="datetime-local"
-                    labelText="Date and Time"
-                    id="dateTime"
-                    name="dateTime"
-                    onBlur={formikProps.handleBlur}
-                    onChange={formikProps.handleChange}
-                    invalid={formikProps.errors.dateTime && formikProps.touched.dateTime}
-                    invalidText={formikProps.errors.dateTime}
-                    value={formikProps.values.dateTime}
-                    min={moment().format(DATETIME_LOCAL_INPUT_FORMAT)}
-                  />
-                </div>
-                <div style={{ width: "23.5rem" }}>
-                  <ComboBox
-                    id="timezone"
-                    initialSelectedItem={formikProps.values.timezone}
-                    //@ts-ignore
-                    items={timezoneOptions}
-                    onChange={({ selectedItem }: { selectedItem: { label: string; value: string } }) => {
-                      const item = selectedItem ?? { label: "", value: "" };
-                      formikProps.setFieldValue("timezone", item);
-                    }}
-                    placeholder="e.g. US/Central (UTC -06:00)"
-                    titleText="Time Zone"
-                  />
-                </div>
-              </>
-            ) : (
-              <CronJobConfig formikProps={formikProps} timezoneOptions={timezoneOptions} />
-            )}
-            {inputs.length ? (
-              <>
-                <p>
-                  <b>Workflow Parameters</b>
-                </p>
-                {inputs}
-              </>
-            ) : null}
-          </ModalBody>
-          <ModalFooter>
-            <Button kind="secondary" onClick={props.modalProps.closeModal}>
-              Cancel
-            </Button>
-            <Button disabled={!formikProps.isValid || props.isLoading} type="submit">
-              {props.type === "create" ? "Create" : "Update"}
-            </Button>
-          </ModalFooter>
-        </ModalForm>
-      )}
-    </DynamicFormik>
   );
 }
